requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

civitai==1.0.0
google-api-python-client==2.161.0
google-auth-oauthlib==1.2.1
pydantic==2.10.6
